Move related components to the dashboard dir
